Arranges javascript codes and adds refresh to comments post request
