+++ .env
@@ -1,3 +1,3 @@
 BASE_MOVIE_API=https://api.tvmaze.com/shows
 BASE_COMMENTS_API=https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/
-COMMENTS_ENDPOINT_ID=dsx7IJ6bdiduLuhqSJgC
+COMMENTS_ENDPOINT_ID=QhwyHRqxtZN5QJVQP2gT
