+++ modules/utils.js
@@ -1,0 +1,11 @@
+// export const refreshButton = () => {
+//   document.querySelectorAll('.add-comments').forEach((button) => {
+//     button.addEventListener('submit', (e) => {
+//       e.preventDefault();
+//       const itemId = button.id[button.id.length - 1];
+//       console.log(itemId);
+//     });
+//   });
+// };
+//
+// export default { refreshButton };
